+++ CHANGELOG.md
@@ -9,6 +9,10 @@
   - lmd-info(1) prints styles info
   - Remove strict warnings on dependencies
   - Do not stream styles and scripts If they are not defined
+  - `banner` property. See [code](examples/features/banner/)
+  - Bugfix #169 If no bundles lmd should not print "bundle" in options
+  - Bugfix #174 apply default value of `sourcemap_www` only for non-strings & replace back slashes with front slashes for source URLs
+  - Test example
 
 **v1.10.x**